backend/requirements.txt
# ERAOTS Backend Dependencies
# Core Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
aiosqlite==0.20.0
alembic==1.13.2
# For PostgreSQL production: pip install asyncpg
# asyncpg==0.29.0

# Cache & Task Queue
redis==5.1.0
celery==5.4.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# Validation & Serialization
pydantic==2.9.1
pydantic-settings==2.5.2
email-validator==2.2.0

# Notifications
httpx==0.27.2

# Export
openpyxl==3.1.5
reportlab==4.2.2

# WebSocket
websockets==12.0

# Utilities
python-dateutil==2.9.0
pytz==2024.1

# Development
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
